third_party/blink/web_tests/virtual/fenced-frame-shadow-dom/external/wpt/html/cross-origin-embedder-policy/anonymous-iframe/fenced-frame-bypass.tentative.https.window-expected.txt

This is a testharness.js-based test.
FAIL fenced-frame-bypass assert_equals: expected "Cannot communicate" but got "Can communicate"
Harness: the test ran to completion.
